Add pivot to UI.Area
Merged with UI.Absolute
